Refactor. Added flow (a js debugger). Renamed app.
