+++ .flowconfig
@@ -6,6 +6,7 @@
 
 # Some modules have their own node_modules with overlap
 .*/node_modules/node-haste/.*
+.*/node_modules/.*
 
 # Ugh
 .*/node_modules/babel.*
@@ -62,4 +63,4 @@ suppress_comment=\\(.\\|\n\\)*\\$FlowIssue\\((\\(>=0\\.\\(2[0-1]\\|1[0-9]\\|[0-9
 suppress_comment=\\(.\\|\n\\)*\\$FlowFixedInNextDeploy
 
 [version]
-0.21.0
+0.22.1
