+++ android/app/build.gradle
@@ -81,7 +81,7 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "com.awesomeproject"
+        applicationId "com.dogshow"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
