+++ android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'AwesomeProject'
+rootProject.name = 'DogShow'
 
 include ':app'
