+++ package.json
@@ -1,13 +1,17 @@
 {
-  "name": "AwesomeProject",
+  "name": "DogShow",
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node_modules/react-native/packager/packager.sh"
+    "start": "node_modules/react-native/packager/packager.sh",
+    "flow": "flow; [[ $?=0||$?=2 ]]"
   },
   "dependencies": {
     "react-native": "^0.21.0",
     "react-redux": "^4.4.1",
     "redux": "^3.3.1"
+  },
+  "devDependencies": {
+    "flow-bin": "^0.22.1"
   }
 }
